20# Product CRUD Process Added

+++ Asp.NetCore10.0_BigData_Analytics_Project/Entities/Product.cs
@@ -5,7 +5,7 @@
         public int ProductID { get; set; }
         public string ProductName { get; set; }
         public string ProductDescription { get; set; }
-        public float UnitPrice { get; set; }
+        public double UnitPrice { get; set; }
         public int StockQuantity { get; set; }
         public int CategoryID { get; set; }
         public Category Category { get; set; }
